Linked to firebase Database
Added libraries and functions to link to firebase database.
